requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyodbc==5.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
Werkzeug==3.1.3
Flask-Login
